model の記述を整理
